#51 update kafka base image to bitnami
Updates the base image for the aissemble-kafka image from wurstmeister
kafka to bitnami/kafka:3.5.  The bitnami image was selected because our
helm chart, aissemble-kafka-chart, is based on the bitnami kafka chart.
The version 3.5 was chosen over the newer 3.7 version image because
there was a major refactor to bitnami's chart in chart version 24.0 that
would have required a significant refactor and testing to our chart to
upgrade beyond. Instead I updated to the last compatible version of the
chart, 23.0.7, which was based on the 3.5 image.