#199 fix Krausening/Quarkus config through helm
Prior to transitioning the Quarkus-based Docker images off of the
`aissemble-quarkus` image, we were setting JAVA_OPTIONS at startup via
the `run-env.sh` script to include the Krausinening configuration
through environment variables.  When it was moved, the script was
dropped and the ENV vars were set directly during the image build.
However, this makes the config unresponsive to the `env` config for
Helm deployments which removes the ability to point Krausening to a
different location at deploy time. For now, we'll just revert to using
`aissemble-quarkus` and the `run-env.sh` script there.  Further work on
Universal Config and a more "serverless" style of deployment will likely
end up illuminating a better answer in the future.