#199 only start Vault during ITs
There is one unit test in `aissemble-extensions-encryption-vault-python`
and 3 ITs. Because of the way the feature files and `environent.py`
setup was structured, the Vault container was spun up even if only the
unit test is being executed and the ITs are excluded.  This causes
problems during the release as the Vault container is not available
during the preparation phase, and generally slows down builds.