Fix Python service containers
While there are several confounding factors, the root issue was a
release by the `bundle` plugin team on Friday. Release 1.4.0 appears to
have a bug so that if a project has an existing virtual environment, the
target location is _ignored_.  This means we weren't creating the venv
bundle at `/opt/venv` as expected.

We weren't seeing this locally because we use in-project virtual
environments that are not copied to CI.  For reasons outside the scope
of this PR, CI was using a poetry.toml file that did _not_ use an
in-project env thus encountering the bug.  To provide better consistency
between CI and local (and between Maven/Habushu and native Poetry) we
will commit `poetry.toml`.