#255 fix PDP image and change Dockerfile delimiter
The rename of the security module in #232 broke the default Policy
Decision Point image, because of the class rename of
`AissembleAttributeExtension` was not propagated to the `pdp.xml` file.
The demo/example policy was also broken because the subject/user name
and the URN of the attribute that the `LocalAttributePoint` class was
looking for was changed, but the test-attribute.json and test-policy.xml
included in the image was not updated. Further, the example JWT token
was for the wrong subject name.