[#366] fix data record python version
We updated the version bound for python to be `>=3.8` in a previous
commit, but missed updating the MDA templates for data record modules.
This resulted in a failure to build projects with data records, as the
version range `^3.11.4` was no longer satisfied when installing the
dependencies of a PySpark pipeline.