[#373] fix chart deployment
Using the `.Chart.AppVersion` for image tags exposed a bug that was
introduced when we moved to opensource.  Previously we were overwriting
path dependencies in helm charts during the release process via the
`release` profile. After open-sourcing our release profile name changed
from `release` to `ossrh-release`, but we didn't update it in the
`extensions-helm` POM.  However, it was incorrect to only rewrite
dependencies during release anyway, as the chart dependencies should be
rewritten _any_ time we're deploying (including snapshots). So instead
we moved the functionality to the `ci` profile (which is also active
during release.)

Also found an issue with release-prepare introduced by #451 preventing
the version update process due to unresolved property.