[#467] document self-hosted runner setup
The documentation for setting up self-hosted runners that can handle the
aissemble build via the Actions Runner Controller project has been added
to `devops`. In addition, the `helm install --dry-run` IT has been
removed and the build action has been adjusted to skip cache-restore
actions if the trigger was a schedule.  The dry run was removed largely
because it just requires too many permissions to function and isn't
adding much more than our strict linting operation and unit tests. The
cache restore skip was added as an alternative to the manual cache
deletion logic we were doing in the build before. General cache cleanup
is handled automatically by S3.