[#469] fix encryption ITs for changed MDA model
Part of the encryption IT scenarios were checking that the right number
of fields were retrieved by the `get_field_list` method.  I'm not sure
that this makes sense as an encryption test.  Either the encryption
method (which uses `get_field_list`) encrypts fields correctly, or it
doesn't. I don't see a ton of value in testing that method specifically
over the actual encryption tests which indirectly test `get_field_list`.

This changeset also splits off the vault scenarios into a separate
feature, as those are the only real "IT" scenarios.  We can't just tag
the scenarios directly because the feature-level tag is used to
determine whether we should stand up the Vault image, and we can't base
this logic at the scenario level because that would tear down and deploy
Vault between each scenario, which is wasteful since we aren't changing
the state of Vault between scenarios.