[#469] use quotes for display version in Antora docs
The 1.10 release caused issues with the version display in our technical
docs. In `antora.yml` we previously weren't quoting the display value,
which in YAML causes type coersion of `1.10` to a floating point number
which is in turn truncated to `1.1`.  The `display_version` field is
always intended to be a string, so it should always be quoted.

This updates our release job to quote the value when updating it and
updates the current `antora.yml` with quotes.