[#788] add migration to specify format in tool.poetry.include
This migration only looks for include lines that are fairly close to the
default that is generated by aiSSEMBLE.  This should cover a large
majority of cases, but for the edge cases a new conditional step has
been added to the upgrade instructions telling projects to account for
the new Poetry 2 behavior manually.

Also fixes the template updates from #779 to account for the fact that
the default project layour is under `src`.